todoList 작업 - 상세보기작업(상세페이지로 이동) - done이 true일 때만 삭제 - done 상태유지

+++ src/components/DailyTodos.vue
@@ -4,9 +4,12 @@
       <button @click="addTodos" class="daily-add">할 일 추가하기</button>
       <div class="daily-list">
         <ul v-for="todo in todos" :key="todo._id">
-          <input type="checkbox" v-model="todo.done" />
+          <input type="checkbox" v-model="todo.done" @click="isDone(todo._id)" />
           <li :class="todo.done ? 'done' : 'notyet'">{{ todo.title }}</li>
-          <img @click="deletTodo(todo._id)" src="../../public/images/png/trash.png" alt="trash" />
+          <div>
+            <button @click="move(todo._id)">상세보기</button>
+            <img @click="deletTodo(todo._id)" src="../../public/images/png/trash.png" alt="trash" />
+          </div>
         </ul>
       </div>
     </section>
@@ -23,9 +26,11 @@
       </div>
       <div class="add">
         <p>Tags</p>
-        <input type="text" v-model="tag" @click="addTags" />
-        <button>추</button>
-        <p v-for="(ii, index) in tags" :key="index">{{ ii }}</p>
+        <div class="tag-box">
+          <input @keypress.enter="addTags" type="text" v-model="tag" />
+          <button @click="addTags">추가</button>
+        </div>
+        <span class="hash-tags" v-for="(ii, index) in tags" :key="index">{{ `# ${ii} ` }}</span>
       </div>
       <button class="todo-add" @click="createTodo">확 인</button>
     </div>
@@ -45,26 +50,52 @@ export default {
     };
   },
   methods: {
+    move(id) {
+      this.$router.push(`/detail/${id}`);
+      this.$store.dispatch('get', id);
+      console.log(this.$store.state.detailData);
+    },
+    isDone(id) {
+      const newTodo = this.todos.filter(el => el._id === id);
+      fetch(`http://localhost:4004/api/todos/${id}`, {
+        method: 'PATCH',
+        headers: {
+          'Content-Type': 'application/json',
+        },
+        body: JSON.stringify({
+          done: !newTodo[0].done,
+        }),
+      });
+    },
     addTags() {
       this.tags.push(this.tag);
+      this.tag = '';
     },
     addTodos() {
       this.addTodo = true;
     },
     closeTodos() {
       this.addTodo = false;
+      this.title = '';
+      this.desc = '';
+      this.tags = [];
     },
     deletTodo(id) {
-      fetch(`http://localhost:4004/api/todos/${id}`, {
-        method: 'DELETE',
-      })
-        .then(res => {
-          console.log(res);
-          this.updateTodo();
+      const newTodo = this.todos.filter(el => el._id === id);
+      if (newTodo[0].done) {
+        fetch(`http://localhost:4004/api/todos/${id}`, {
+          method: 'DELETE',
         })
-        .catch(e => {
-          console.log(e);
-        });
+          .then(res => {
+            console.log(res);
+            this.updateTodo();
+          })
+          .catch(e => {
+            console.log(e);
+          });
+      } else {
+        alert('할 일은 하고 삭제하세요');
+      }
     },
     updateTodo() {
       fetch('http://localhost:4004/api/todos')
@@ -76,22 +107,34 @@ export default {
         });
     },
     createTodo() {
-      fetch('http://localhost:4004/api/todos', {
-        method: 'POST',
-        body: {
-          title: this.title,
-          desc: this.desc,
-          tags: this.tags,
-        },
-      })
-        .then(res => {
-          console.log(res);
-          this.updateTodo();
+      if (this.title.length > 0 && this.desc.length > 0 && this.tags.length > 0) {
+        fetch('http://localhost:4004/api/todos', {
+          method: 'POST',
+          headers: {
+            'Content-Type': 'application/json',
+          },
+          body: JSON.stringify({
+            title: this.title,
+            desc: this.desc,
+            tags: this.tags,
+          }),
         })
-        .catch(e => {
-          console.log(e);
-        });
+          .then(() => {
+            this.updateTodo();
+          })
+          .catch(e => {
+            console.log(e);
+          });
+        this.closeTodos();
+      } else {
+        alert('내용을 입력하세요');
+      }
     },
+  },
+  clearInfo() {
+    this.title = '';
+    this.desc = '';
+    this.tags = [];
   },
   created() {
     fetch('http://localhost:4004/api/todos')
